+++ src/main/java/com/example/OskrApplication.java
@@ -2,16 +2,12 @@ package com.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
 
-@SpringBootApplication(exclude = SecurityAutoConfiguration.class) // TODO 一時的にSpringSecurityによる認証を無効化
-//@SpringBootApplication
+//@SpringBootApplication(exclude = SecurityAutoConfiguration.class) // TODO 一時的にSpringSecurityによる認証を無効化
+@SpringBootApplication
 public class OskrApplication {
 
 	public static void main(String[] args) {
 		SpringApplication.run(OskrApplication.class, args);
 	}
-
-
-
 }
