+++ src/main/resources/templates/filmworks/filmworks.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="ja"  xmlns:th="http://www.thymeleaf.org"
+				 xmlns:sec="http://www.thymeleaf.org/extras/spring-security"
 	th:replace="~{layout/base :: layout( ~{},~{::body/content()})}">
   <head>
     <meta charset="UTF-8" />
@@ -75,7 +76,9 @@
 	      </section>
 	      <!-- ▲▲▲ 作品情報セクション ▲▲▲ -->
 	      <!-- ▼▼▼ レビュー投稿セクション ▼▼▼ -->
+	      <p class="text-center">ユーザ名：<span sec:authentication="principal.user.name"></span></p>
 	      <section class="py-4 bg-light border-top border-bottom">
+<!--	      <th:section th:if="${sec:authentication != null}"> -->
 	        <!-- レビュー コンテンツ行 -->
 	        <div class="col-12 mb-1">
 	        	<h2 class="fw-bold text-center">
@@ -128,7 +131,8 @@
 	            </div>
 	          </form>
 	        </div>
-	      </section>
+	        </section>
+<!--	      </th:section> -->
 	      <!-- ▼▼▼ 作品レビューセクション ▼▼▼ -->
 	      <section class="py-2 bg-light border-top border-bottom">
 	        <!-- コンテナ -->
@@ -178,6 +182,7 @@
 	                      <p th:text="${review.content}">
 	                        前半は良かったけど途中から思ってた展開とずれてきて自分の感情との齟齬が最後まで埋められなかったなぁ。種明かしシーンが長すぎて。
 	                      </p>
+<!--	                      <th:block th:if="${sec:authentication != null}"> -->
 	                      <div class="mb-1 mt-2">
 	        				<div class="fw-bold text-center">
 	        					<span class="d-block text-secondary text-muted middle">POST REPLY</span>
@@ -194,6 +199,7 @@
 	                      		</div>
 	                      	</div>
 	                      </form>
+<!--	                      </th:block> -->
 	                      <!-- ネタばれ無しタブ リプライ 表示エリア -->
 	                      <div class="mb-2"></div>
 	                      <div class="mb-1 mt-0">
