+++ src/main/resources/templates/login/loginForm.html
@@ -73,7 +73,7 @@
 	                	</div>
 	            	</form>
 	        	<button type="button" class="btn btn-primary btn-lg" data-bs-toggle="modal" data-bs-target="#modal_wrap2">新規登録</button>
- 			        <form method="post" th:action="@{/signup}" th:object="${signupForm}">
+ 			        <form method="post" th:action="@{/login/signup}" th:object="${signupForm}">
 	                	<div class="modal fade" id="modal_wrap2" tabindex="-1" data-bs-backdrop="static">
 	                    	<div class="modal-dialog modal-dialog-centered">
 	                        	<div class="modal-content">
